chore: improve NoSsr exports
