+++ packages/chakra-ui-core/src/NoSsr/index.js
@@ -1,42 +1,2 @@
-/**
- * This component is a modification of the fine work of @egoist
- * @see https://github.com/egoist/vue-client-only/blob/master/src/index.js
- */
-const NoSsr = {
-  name: 'NoSsr',
-  functional: true,
-  props: {
-    placeholder: String,
-    placeholderTag: {
-      type: String,
-      default: 'div'
-    }
-  },
-  render (h, { parent, slots, props }) {
-    const { default: defaultSlot = [], placeholder: placeholderSlot } = slots()
-
-    if (parent._isMounted) {
-      return defaultSlot
-    }
-
-    parent.$once('hook:mounted', () => {
-      parent.$forceUpdate()
-    })
-
-    if (props.placeholderTag && (props.placeholder || placeholderSlot)) {
-      return h(
-        props.placeholderTag,
-        {
-          class: ['client-only-placeholder']
-        },
-        props.placeholder || placeholderSlot
-      )
-    }
-
-    // Return a placeholder element for each child in the default slot
-    // Or if no children return a single placeholder
-    return defaultSlot.length > 0 ? defaultSlot.map(() => h(false)) : h(false)
-  }
-}
-
+import NoSsr from './NoSsr'
 export default NoSsr
